Adding completed Tesla asset Blender
This is an asset for the Tesla Model S, based on a tutorial by YouTuber Blue Inversion. Note, this tutorial is meant to create a Tesla asset render in a showroom environment. As this was my first experience using Blender, the tutorial has been followed from start to finish (with the exclusion of some scenes that would be additional work) with some minor aesthetic changes to move in line with ultimate vision for these assets as they will exist in game.

Future changes to this file are expected when time to import into Unreal Engine 5